Implements Pundit Policies for Wikis and Articles
In the Wiki model, we have the following scopes:

:all_viewable (returns all Wikis, sorted by Title)
:publicly_viewable (returns Public Wikis, sorted by Title)
:privately_viewable (returns Public Wikis and Private Wikis that the
User created, sorted by Private then Title)
:owned_by (returns Wikis created by the user)

These scopes are used to enforce Pundit policies.

In our application security model:

Admins can perform CRUD operations on all Wiki and Article types
Premium users can CRU all Article Types and D Private Wikis they have
created
Standard users can CRU all Public Articles

+++ app/controllers/wikis_controller.rb
@@ -2,18 +2,22 @@ class WikisController < ApplicationController
   before_action :load_wiki, only: [:show, :edit, :update, :destroy]
 
   def index
-    @wikis = Wiki.all
+    @wikis = Wiki.visible_to(current_user)
+    authorize @wikis, @articles
   end
 
   def show
+    authorize @wiki, @article
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki, @article
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
+    authorize @wiki, @article
     if @wiki.save
       redirect_to @wiki, notice: "Wiki was created successfully."
     else
@@ -23,20 +27,25 @@ class WikisController < ApplicationController
   end
 
   def edit
+    authorize @wiki, @article
   end
 
   def update
     @wiki.assign_attributes(wiki_params)
+    authorize @wiki, @article
 
     if @wiki.save
       redirect_to @wiki, notice: "Wiki was updated successfully."
     else
       flash[:error] = "Error updating wiki. Please try again."
       render :new
-    end    
+    end 
+       
   end
 
   def destroy
+    authorize @wiki, @article
+
     if @wiki.destroy
       redirect_to @wiki, notice: "Wiki was deleted successfully."
     else
